DLK-149 classes to standardise requestIDs and sessionIDs for batches
It is important that associated audit events can be tied together with
requestID even in the case of scheduled tasks.

Frontend HTTP requests are given a requestID which is then added to the
HeaderCarrier so that all subsequent calls forward the requestID for use
in audits.

This code provideds a standardised way to create a HeaderCarrier
for use in batch processes so that any http calls the batch service makes
pass on a requestID and sessionID for use in any audit events created by
those services.